Sort decisions by integer value of id

+++ src/main/kotlin/nl/avisi/structurizr/site/generatr/site/model/DecisionsTableViewModel.kt
@@ -7,7 +7,7 @@ fun PageViewModel.createDecisionsTableViewModel(decisions: Collection<Decision>,
     TableViewModel.create {
         headerRow(headerCell("ID"), headerCell("Date"), headerCell("Status"), headerCell("Title"))
         decisions
-            .sortedBy { it.id }
+            .sortedBy { it.id.toInt() }
             .forEach {
                 bodyRow(
                     headerCell(it.id),
